# shim/openapi.yaml
openapi: 3.0.3
info:
  title: Trasi — shim (contratto v0, congelato in B0 · gate V-09)
  version: 0.0.0
  description: "Contratto d'integrazione fra Onyx e lo shim Trasi, congelato nel blocco B0 (gate V-09) per rompere il ciclo\
    \ B2↔B3:\nB2 (SA-Onyx) registra questo documento come tool custom **prima** che gli endpoint siano implementati; B3 (SA-Shim)\n\
    implementa le stesse `operationId`, senza cambiarle.\n\nRegole di contratto (non negoziabili):\n- **Un solo server**:\
    \ Onyx sostituisce lato server il letterale `USER_EMAIL` con l'email dell'utente autenticato e\n  invia l'header `X-Onyx-User-Email`;\
    \ lo shim risolve l'identità in `identita_onyx` e apre `BEGIN; SET LOCAL ROLE <ruolo_db>; …; COMMIT`.\n  La RLS del database\
    \ è l'unica autorità sui permessi.\n- **Nessun dato personale**: nessuno schema ammette campi liberi (`additionalProperties:\
    \ false` ricorsivo) e nessun\n  campo è destinato al cittadino. `motivazione` e `payload` sono filtrati dallo shim (email,\
    \ telefono italiano,\n  codice fiscale) → 422 `dato_personale_sospetto`.\n- **Errori**: 401 chiave shim assente o errata\
    \ · 403 identità Onyx non riconosciuta o RLS che nega\n  (`{\"detail\":\"da approvare in coda\"}`) · 409 proposta scaduta\
    \ · 422 payload o parametri non ammessi · 501 non ancora\n  implementato (stub v0). I corpi d'errore usano l'involucro\
...
    \ da B3-SHM-07 dell'architettura.\n"
servers:
- url: http://shim:8000/v1/u/USER_EMAIL
...
paths:
...
  /vicino_a:
    get:
      operationId: vicino_a
      tags:
      - kb
      - esterno
      summary: 'Trova i luoghi di un certo tipo vicino a una Casa di Quartiere, unendo la memoria della rete (KB) e i punti
        di interesse esterni di OpenStreetMap. Usalo per ogni domanda di vicinanza ("il bar più vicino", "una farmacia vicino
        a Tuturano", "un CAF aperto adesso"). Ogni risultato dichiara se è verificato (KB) o esterno e porta un badge già
        formattato: riportalo così com''è. Se una fonte esterna non risponde, dichiara `fonti_esterne[].stato` senza inventare
        dati.'
      parameters:
      - name: casa
        in: query
        required: false
        description: Slug della Casa di Quartiere. Se omesso, si usa la Casa dell'operatore autenticato (dall'identità); è
          obbligatorio solo per i ruoli senza Casa (es. rete).
        schema:
          type: string
          minLength: 1
      - name: tipo
        in: query
        required: true
        description: Tipo di luogo cercato (obbligatorio); mappa chiusa verso i tag OpenStreetMap. Un valore fuori elenco
          risponde 422 con l'elenco ammesso.
        schema:
          type: string
          enum:
          - bar
          - farmacia
          - caf
          - fermata
          - poste
          - medico
          - ospedale
          - supermercato
          - biblioteca
          - parco
          - banca
          - comune
      - name: aperto_adesso
        in: query
        required: false
        description: 'Se `true`, esclude i luoghi noti come chiusi in questo momento e ordina prima gli aperti. I luoghi senza
          orari noti **non vengono scartati**: restano in coda con `aperto_adesso: null` e `orari_nota: "orari non disponibili"`.'
        schema:
          type: boolean
      - name: raggio_m
        in: query
        required: false
        description: Raggio di ricerca in metri, facoltativo; se omesso si usa `casa.raggio_m` o il parametro `raggio_vicinanza_m`.
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Luoghi KB ed esterni entro il raggio, con provenienza, fiducia e badge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaVicinoA'
        '401':
          $ref: '#/components/responses/ChiaveShimNonValida'
        '403':
          $ref: '#/components/responses/IdentitaNonRiconosciuta'
        '422':
          $ref: '#/components/responses/ParametriNonAmmessi'
        '501':
          $ref: '#/components/responses/NonImplementato'
...
components:
  responses:
    ChiaveShimNonValida:
      description: La chiave dello shim (header `X-Trasi-Key`) è assente o errata.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    IdentitaNonRiconosciuta:
      description: L'email dell'operatore non è in `identita_onyx` (403, nessuna query eseguita), oppure la RLS nega l'operazione
        (`{"detail":"da approvare in coda"}`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
...
    ParametriNonAmmessi:
      description: 'Parametri o corpo non ammessi: campo non previsto (`additionalProperties: false`), parametro di query
        non dichiarato dall''operazione (l''errore elenca quelli sconosciuti e quelli ammessi), valore fuori vocabolario,
        `motivazione` oltre 80 caratteri, destinazione mancante per `inviata_altrove`, oppure dato personale sospetto (email,
        telefono italiano, codice fiscale) → `dato_personale_sospetto`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    NonImplementato:
      description: Operazione prevista dal contratto v0 e non ancora implementata (stub 501 di B0). L'implementazione reale
        è il blocco B3, senza cambiare `operationId`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
  schemas:
    Errore:
      type: object
      additionalProperties: false
      required:
      - detail
      description: Involucro unico degli errori; il dettaglio è una stringa in italiano, adatta anche al LLM.
      properties:
        detail:
          type: string
          description: Spiegazione dell'errore (es. «da approvare in coda», «non implementato (contratto v0)», «dato_personale_sospetto»).
    FonteEsterna:
      type: object
      additionalProperties: false
      required:
      - fonte
      - stato
      - ms
      description: Esito del tentativo di interrogare una fonte esterna. Non solleva mai eccezione al chiamante.
      properties:
        fonte:
          type: string
          description: Nome della fonte esterna interrogata (es. OpenStreetMap, SearXNG).
        stato:
          type: string
          enum:
          - ok
          - timeout
          - errore
          - scartata_fiducia
          description: '`ok` risposta utilizzata · `timeout` oltre il tempo dichiarato · `errore` fonte non raggiungibile
            o risposta illeggibile · `scartata_fiducia` risultati sotto la soglia `fiducia_min_esterna`.'
        ms:
          type: integer
          minimum: 0
          description: Durata del tentativo in millisecondi.
    ItemVicinanza:
      type: object
      additionalProperties: false
      required:
      - provenienza
      - nome
      - tipo
      - indirizzo
      - lat
      - lon
      - distanza_m
      - aperto_adesso
      - orari_testo
      - orari_nota
      - fonte
      - url
      - data_aggiornamento
      - fiducia
      - consultato_ts
      - badge
      description: Un luogo vicino alla Casa, dalla memoria della rete (kb) o da una fonte esterna (esterna). I campi facoltativi
        sono `null`, non omessi, così il LLM vede sempre la stessa forma.
      properties:
        provenienza:
          type: string
          enum:
          - kb
          - esterna
          description: Da dove viene il dato; determina il testo del badge.
        nome:
          type: string
          description: Nome del luogo.
        tipo:
          type: string
          description: Tipo richiesto (es. bar, farmacia).
        indirizzo:
          type: string
          description: Indirizzo del luogo; stringa vuota se la fonte non lo fornisce.
        lat:
          type: number
          minimum: -90
          maximum: 90
          description: Latitudine in gradi decimali (WGS84).
        lon:
          type: number
          minimum: -180
          maximum: 180
          description: Longitudine in gradi decimali (WGS84).
        distanza_m:
          type: number
          minimum: 0
          description: Distanza in metri dalla Casa di riferimento.
        aperto_adesso:
          type: boolean
          nullable: true
          description: '`true` se risulta aperto adesso, `false` se risulta chiuso, `null` se non si conoscono gli orari (caso
            frequente nei dati OpenStreetMap).'
        orari_testo:
          type: string
          nullable: true
          description: Orari leggibili per una persona (es. «lun-ven 9:00-13:00»); `null` se non noti.
        orari_nota:
          type: string
          nullable: true
          description: Nota sugli orari quando mancano (es. «orari non disponibili»); `null` quando gli orari sono noti.
        fonte:
          type: string
          description: Fonte del dato (es. «Comune di Brindisi», «OpenStreetMap»).
        url:
          type: string
          format: uri
          nullable: true
          description: Collegamento alla fonte o al dato originale; `null` se non disponibile.
        data_aggiornamento:
          type: string
          format: date
          nullable: true
          description: Data dell'ultimo aggiornamento del dato in memoria; `null` per i dati esterni.
        fiducia:
          type: integer
          minimum: 1
          maximum: 3
          description: Livello di affidabilità della fonte (1 basso, 3 alto).
        consultato_ts:
          type: string
          format: date-time
          description: Istante in cui il dato è stato consultato o letto.
        badge:
          type: string
          description: Etichetta di provenienza già formattata, da riportare nel testo così com'è, senza ricomporla (es. «[KB
            · Comune di Brindisi · agg. 10/09/2026 · affidabilità 3]»).
...
    RispostaVicinoA:
      type: object
      additionalProperties: false
      required:
      - casa
      - tipo
      - raggio_m
      - items
      - fonti_esterne
      description: Luoghi vicini alla Casa, ordinati per gruppo (prima `kb`, poi `esterna`), poi per apertura (`true`, `null`)
        e infine per distanza.
      properties:
        casa:
          type: string
          description: Slug della Casa da cui è stata misurata la distanza.
        tipo:
          type: string
          description: Tipo di luogo cercato.
        raggio_m:
          type: integer
          minimum: 1
          description: Raggio effettivamente applicato, in metri.
        items:
          type: array
          description: Luoghi trovati entro il raggio, dalla memoria della rete e dalle fonti esterne.
          items:
            $ref: '#/components/schemas/ItemVicinanza'
        fonti_esterne:
          type: array
          description: Esito di ogni fonte esterna interrogata (anche quando non ha risposto).
          items:
            $ref: '#/components/schemas/FonteEsterna'
...